Fix up linting, and add consistency
