+++ lib/mods/clean-console.js
@@ -1,4 +1,6 @@
-const {getConsoleStatments} = require('./utils.js');
+'use strict';
+
+const {getConsoleStatements} = require('./utils.js');
 
 /**
  * Replaces console statements without ExpressionStatement parents
@@ -7,12 +9,13 @@ const {getConsoleStatments} = require('./utils.js');
  * @param {any} j - the jscodemod api
  * @param {Collection} root - the root Collection
  *
+ * @return {Collection} - the modified root Collection
+ *
  * @example
  * cleanConsoleExpressions(j, root);
  */
-
-const cleanConsoleExpressions = (j, root) => {
-  getConsoleStatments(j, root).filter((p) => {
+function cleanConsoleExpressions(j, root) {
+  getConsoleStatements(j, root).filter((p) => {
     const parentType = p.parentPath.value.type;
 
     return parentType !== 'ExpressionStatement';
@@ -30,12 +33,13 @@ const cleanConsoleExpressions = (j, root) => {
  * @param {any} j - the jscodemod api
  * @param {Collection} root - the root Collection
  *
+ * @return {Collection} - the modified root Collection
+ *
  * @example
  * cleanConsoleStatements(j, root);
  */
-
-const cleanConsoleStatements = (j, root) => {
-  getConsoleStatments(j, root).forEach((path) => {
+function cleanConsoleStatements(j, root) {
+  getConsoleStatements(j, root).forEach((path) => {
     if (path.parentPath.value.type === 'ExpressionStatement') {
       j(path.parentPath).replaceWith(() => {
         return;
@@ -52,7 +56,7 @@ const cleanConsoleStatements = (j, root) => {
  * @param {any} j - the jscodemod api
  * @param {Collection} root - the root Collection
  */
-const cleanConsole = (j, root) => {
+function cleanConsole(j, root) {
   cleanConsoleExpressions(j, root);
   cleanConsoleStatements(j, root);
 };
