+++ lib/mods/utils.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * Gets the Collection of CallExpressions with callee name console.
  *
@@ -6,10 +8,9 @@
  * @return {Collection} - a Collection of console statements
  *
  * @example
- * getConsoleStatments(j, root).filter(...);
+ * getConsoleStatements(j, root).filter(...);
  */
-
-const getConsoleStatments = (j, root) => {
+function getConsoleStatements(j, root) {
   return root.find(j.CallExpression, {
     callee: {
       object: {
@@ -20,5 +21,5 @@ const getConsoleStatments = (j, root) => {
 };
 
 module.exports = {
-  getConsoleStatments,
+  getConsoleStatements,
 };
