+++ lib/rewriter.js
@@ -1,6 +1,16 @@
+'use strict';
+
 const cleanConsole = require('./mods/clean-console.js');
 
-module.exports = (fileInfo, api, options) => {
+/**
+ * description
+ *
+ * @param {File} fileInfo - the file to run the rewriter on
+ * @param {any} api - the API containing jscodeshift
+ *
+ * @return {String} the source of the newly rewritten file
+ */
+function rewriter(fileInfo, api) {
   const j = api.jscodeshift;
   const root = j(fileInfo.source);
 
@@ -8,3 +18,5 @@ module.exports = (fileInfo, api, options) => {
 
   return root.toSource();
 };
+
+module.exports = rewriter;
